refactor: extract shared post-processing and prompt-building for improve pipeline
Extract duplicated logic between improve.ts and batch-improve.ts into
shared modules to improve maintainability:

- postProcessImproveResult() in post-process.ts: MDX extraction from code
  fences, JSON-wrapped response detection, validation, truncation guard,
  lastEdited update, frontmatter repair, entity link cleanup, slug
  conversion, and footnote conversion.

- buildImproveContext() in build-context.ts: entity lookup, KB facts
  loading, template resolution, objectivity context, and IMPROVE_PROMPT
  assembly.

- Export COST_MAP and BATCH_COST_MAP from page-router.ts and import in
  orchestrator.ts instead of hardcoding.

Follow-up to #2924 which identified these DRY issues during review.
